Fixed an error in FinnhubAPI.GetPreviousClose

+++ TurtleAPI/PolygonIO/PolygonAPI.cs
@@ -58,7 +58,7 @@ namespace TurtleAPI.PolygonIO
         private static DateTime ParseUnixTimestamp(decimal t)
         {
             var epoch = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
-            var dateTime = epoch.AddMilliseconds((float)t);
+            var dateTime = epoch.AddMilliseconds((double)t);
             return dateTime;
         }
         public static IEnumerable<MarketHoliday>? GetMarketHoliday()
